Hack: Disable CERT_E_REVOCATION_FAILURE check
This prevents the authentication TLS connection from being terminated on
our end, which triggers a race in Nvidia's servers that can end with us
receiving a 0xC0F52132 error.

Note that while it is not normally advisable to delete security checks,
the likelihood of the revocation list check doing anything is
practically nil. Until we fix it properly, we can just disable it with
no ill effect.

Signed-off-by: Richard Yao <ryao@gentoo.org>